Fix a bunch of CMake compiler warnings about implicit data conversion
Bug: 436300898
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256037
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>

+++ src/dawn/native/d3d12/BindGroupD3D12.cpp
@@ -109,7 +109,7 @@ MaybeError BindGroup::InitializeImpl() {
                         // byte aligned. Since binding.size and binding.offset are in bytes,
                         // we need to divide by 4 to obtain the element size.
                         D3D12_UNORDERED_ACCESS_VIEW_DESC desc;
-                        desc.Buffer.NumElements = binding.size / 4;
+                        desc.Buffer.NumElements = static_cast<uint32_t>(binding.size / 4);
                         desc.Format = DXGI_FORMAT_R32_TYPELESS;
                         desc.ViewDimension = D3D12_UAV_DIMENSION_BUFFER;
                         desc.Buffer.FirstElement = binding.offset / 4;
@@ -134,7 +134,7 @@ MaybeError BindGroup::InitializeImpl() {
                         desc.ViewDimension = D3D12_SRV_DIMENSION_BUFFER;
                         desc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
                         desc.Buffer.FirstElement = binding.offset / 4;
-                        desc.Buffer.NumElements = binding.size / 4;
+                        desc.Buffer.NumElements = static_cast<uint32_t>(binding.size / 4);
                         desc.Buffer.StructureByteStride = 0;
                         desc.Buffer.Flags = D3D12_BUFFER_SRV_FLAG_RAW;
                         d3d12Device->CreateShaderResourceView(
@@ -217,7 +217,7 @@ MaybeError BindGroup::InitializeImpl() {
          ++bindingIndex) {
         if (bgl->IsStorageBufferBinding(bindingIndex)) {
             mDynamicStorageBufferLengths[dynamicStorageBufferIndex++] =
-                GetBindingAsBufferBinding(bindingIndex).size;
+                static_cast<uint32_t>(GetBindingAsBufferBinding(bindingIndex).size);
         }
     }
