+++ src/dawn/native/d3d12/ResourceAllocatorManagerD3D12.cpp
@@ -241,8 +241,8 @@ uint32_t ComputeExtraArraySizeForIntelGen12(uint32_t width,
 
     // Texture should be aligned on both tile width (512 bytes) and tile height (128 rows) on Intel
     // Gen12 GPU
-    uint32_t mainTileCols = Align(totalWidth, tileWidth) / tileWidth;
-    uint32_t mainTileRows = Align(totalHeight, tileHeight) / tileHeight;
+    uint64_t mainTileCols = Align(totalWidth, tileWidth) / tileWidth;
+    uint64_t mainTileRows = Align(totalHeight, tileHeight) / tileHeight;
     uint64_t mainTileCount = mainTileCols * mainTileRows;
 
     // There is a bug in Intel old drivers to compute the auxiliary memory size (auxSize) of the
